+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+- Restored the `--verbose-process-command` and `--verbose-rerun-command` options, removed in version 4.0.0.
+See [#952](https://github.com/grosser/parallel_tests/pull/952).
+`--verbose-command` continues to be supported and is equivalent to set the 2 options above.
 
 ### Breaking Changes